able to show the amount spent on tags and merchants

+++ db/spending_tracker.sql
@@ -17,5 +17,5 @@ CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
     merchant_id INT REFERENCES merchants(id) ON DELETE CASCADE,
     tag_id INT REFERENCES tags(id) ON DELETE CASCADE,
-    amount FLOAT
+    amount DECIMAL(5,2)
 );
